adjustments and big run before running on solstr
