+++ .bash_profile
@@ -1,0 +1,1 @@
+export PYTHONPATH="$PWD"
