+++ analysis/train_value_function.py
@@ -32,8 +32,8 @@ def train_value_function(
 
 
 if __name__ == "__main__":
-    SAMPLE_SIZE = 100
-    NUMBER_OF_CLUSTERS = 10
+    SAMPLE_SIZE = 2500
+    NUMBER_OF_CLUSTERS = 50
 
     POLICY = decision.RolloutValueFunctionPolicy(
         decision.EpsilonGreedyValueFunctionPolicy(
